requirements.txt
langchain
langchain-openai
python-dotenv
langchain_community
langchain_core
streamlit
